requirements.txt
ExifRead==3.0.0
pytest==7.4.3
pytest-cov==4.1.0
Pillow==10.2.0
piexif==1.1.3
